refactor: remove unused code & adapt tests
