+++ lib/features/journal/model/entry_state.dart
