+++ lib/features/journal/model/entry_state.freezed.dart
