+++ lib/features/tasks/ui/task_app_bar.dart
@@ -46,12 +46,7 @@ class TaskSliverAppBar extends ConsumerWidget {
       titleSpacing: 0,
       scrolledUnderElevation: 10,
       elevation: 10,
-      title: Column(
-        children: [
-          LinkedDuration(task: item),
-          LinkedDuration2(taskId: item.id),
-        ],
-      ),
+      title: LinkedDuration(taskId: item.id),
       bottom: PreferredSize(
         preferredSize: const Size.fromHeight(25),
         child: AnimatedOpacity(
