+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: Achieve your goals and keep your data private with Lotti.
 publish_to: 'none'
-version: 0.9.537+2758
+version: 0.9.538+2759
 
 msix_config:
   display_name: LottiApp
